Fix the NettyOioServer example
The code example for `NettyNioServer` doesn't work, because the server
doesn't flush a message wrote to a client and closes the TCP connection
immediately. You can verify that by running a server on a TCP port and
connecting to it via telnet.

The fix is to use the `writeAndFlush` method as described in the book.